Pay Check Estimator
This is an application that I have created to help hourly employees to get an estimate of how much their pay check will be for the bi-weekly pay period.  The numbers for the state tax is based on Louisiana income tax branch with the income of $50,000 per year salary.  I am looking for feed back on how I can make this application better so that I may be able to turn it into an android and/or iOS app.